rename Register component to SignUp
